+++ src/ComplexityMeasures.jl
@@ -1,11 +1,11 @@
-module Entropies
+module ComplexityMeasures
 
 # Use the README as the module docs
 @doc let
   path = joinpath(dirname(@__DIR__), "README.md")
   include_dependency(path)
   read(path, String)
-end Entropies
+end ComplexityMeasures
 
 using Reexport
 @reexport using StateSpaceSets
@@ -32,19 +32,26 @@ include("complexity_measures/complexity_measures.jl")
 
 # Update messages:
 using Scratch
-display_update = true
-version_number = "2.0.0"
+display_update = false
+version_number = "2.0"
 update_name = "update_v$(version_number)"
 update_message = """
-\nUpdate message: Entropies v$(version_number)\n
-- An overall overhaul of the documentation and API of Entropies.jl has been performed.
-- A huge amount of new content has been added, which is best seen by visiting the
-  the online documentation. Some examples are Tsallis entropy and spatial permutation
-  entropy, and much more.
-- In summary, all entropies and normalized entropies are computing using the
-  `entropy` and `entropy_normalized` functions, which dispatch on entropy types such
-  as `Renyi()`, `Shannon()` or `Tsallis()`.
-- New constructors for discretizing: `OrdinalPatternEncoding, GaussianCDFEncoding`.
+Update message: ComplexityMeasures.jl v$(version_number)!
+
+Entropies.jl has been completely overhauled,
+and has been renamed to ComplexityMeasures.jl.
+
+Along with the overhaul comes a massive amount of new features, an entirely new API,
+extendable and educative code, dedicated documentation pages, and more!
+At the moment we estimate we offer at least 90 unique quantities characterizing
+complexity in the context of nonlinear dynamics and complex systems.
+We believe it is best to learn all of this by visiting the online documentation!
+
+We tried our best to keep pre-2.0 functions working and throw deprecation warnings.
+If we missed code that should be working, please let us know by opening an issue.
+
+For example, `genentropy(x::Array_or_Dataset, ε::Real; q, base)` is deprecated
+in favor of `entropy(Renyi(q, base), ValueHistogram(ε), x)`.
 """
 
 if display_update