+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # CHANGELOG
 
-Changelog is kept with respect to version 0.11 of Entropies.jl.
+Changelog is kept with respect to version 0.11 of Entropies.jl. From version v2.0 onwards, this package has been renamed to ComplexityMeasures.jl.
+
+## 2.1
+
+- Added `Gao` estimator for differential Shannon entropy.
 
 ## 2.0
 
-The API for Entropies.jl has been completely overhauled,
-but along with the overhaul comes a massive amount of new features.
+The API for Entropies.jl has been completely overhauled, and the package has been renamed to ComplexityMeasures.jl.
+Along with the overhaul comes a massive amount of new features, an entirely new API, extendable and educative code, dedicated documentation pages, and more!
+
 We believe it is best to learn all of this by visiting the online documentation.
 
 We tried our best to keep pre-2.0 functions working and throw deprecation warnings.